Updated project based on feedback
