--- ddd/src/main/java/com/baeldung/dddhexagonalapp/coreapp/service/CardHolderServiceImpl.java
@@ -1,31 +1,0 @@
-package com.baeldung.dddhexagonalapp.coreapp.service;
-
-import java.util.List;
-import java.util.Optional;
-
-import com.baeldung.dddhexagonalapp.coreapp.repository.CardHolderRepository;
-import com.baeldung.dddhexagonalapp.coreapp.domain.CardHolder;
-
-public class CardHolderServiceImpl implements CardHolderService {
-
-    private CardHolderRepository repo;
-
-    public CardHolderServiceImpl(CardHolderRepository respository) {
-        repo = respository;
-    }
-
-    public CardHolder registerCardHolder(CardHolder cardHolder) {
-
-        return repo.CreateCardHolder(cardHolder);
-
-    }
-
-    public List<CardHolder> getAllCardHolders() {
-        return repo.findAll();
-    }
-
-    public Optional<CardHolder> findCardHolderById(int cardHolderId) {
-        return repo.findCardHolderById(cardHolderId);
-    }
-
-}
